Create README.md
Computing Technology in Society - Project Base Mock-up Site.  Site contains index page for the site, images, CSS styles and miscellaneous pages.  Essentially a site for the project, hosting the different aspects of the profile of the team's chosen technology or topic and the supporting assignment activities in the process support pages.